requirements.txt
anthropic==0.69.0
Django==5.2
django_allauth==65.7.0
google_api_python_client==2.169.0
protobuf==6.32.1
python-dotenv==1.1.1
twilio==9.3.7
gunicorn
dj-database-url
psycopg2-binary
whitenoise
cryptography
